Runtime Polymorphism: Basic document_t: value types extended
Do your own memory management - don't create garbage for your client
to clean-up.
The semantics of copy are to create a new object which is equal to, and
logically disjoint from, the original.
Copy constructor must copy the object. The compiler is free to elide
copies so if the copy constructor does something else the code is
incorrect.
When a type manages _remote_ parts it is necessary to supply a copy
constructor.
 * If you can, use an existing class (such as a vector) to manage remote
 parts.

Assignment is consistent with copy. Generally:
 T x; x = y; is equivalent to T x = y;
Assignment satisfying the _strong exception guarantee_ is a nice
property.
 * Either complete successfully or throw an exception, leaving the
 object unchanged.
Assignment (like all other operations) must satisfy the basic exception
guarantee.
Don't optimise for rare cases which impact common cases.
 * Don't test for self-assignment to avoid the copy.

Guideline: The Private Implementation (Pimpl), or Handle-Body, idion is
good for separating the implementation and reducing compile times.